Audit Log: Skeleton for D-Bus AuditLog interface
Provides structure but no function yet for D-Bus interfaces.
Added parseAuditLog method for debug purposes since GetAuditLog method
is causing a core dump. Seems to be a problem with the way the unix_fd
return is being handled.

Tested:
Started service in background and was able to call each method and see
the debug message.

Signed-off-by: Janet Adkins <janeta@us.ibm.com>